crowdstrike-mssp/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
cachetools==5.5.0
python-jose[cryptography]==3.3.0
slowapi==0.1.9
structlog==24.4.0
asyncio-throttle==1.0.2
